\name{anaBehav}
\alias{anaBehav}
\title{Compute some behavioral variables from TDR dataset}
\usage{
anaBehav(tdr, dvs)
}
\arguments{
  \item{tdr}{The TDR dataset.}

  \item{dvs}{Optional. A table with dives/surfaces/bootoms
  indices as returned by \code{\link{divesID}} or
  \code{\link{anaDives}}.}
}
\description{
Compute some behavioral variables from TDR dataset
}

