\name{as.ses.list}
\alias{as.ses}
\alias{as.ses.list}
\title{as.ses.list}
\usage{
  as.ses.list(x, type = c("classic", "3D"))
}
\arguments{
  \item{x}{The object to coerce to 'ses'}

  \item{type}{The type of 'ses' desired. To choose in
  \code{c("classic", "3D")}.}
}
\description{
  S3 method of \code{as.ses} for list objects.
}

