\docType{data}
\name{formatSES}
\alias{formatSES}
\title{SES import preferences}
\format{Two data frames, one for each of TDR and statdives datasets, storing the names to give to the initial matlab variable (alias) and whether or not they should be imported from the .mat file (keep).}
\usage{
  tdrfmt
}
\description{
  User preferences about SES importation.
}
\keyword{datasets}

