\name{is.SES}
\alias{is.SES}
\title{is.SES}
\usage{
is.SES(x)
}
\arguments{
  \item{x}{The object}
}
\description{
Test if an object belongs to the family of SES package
object.
}
\seealso{
Other is.thing: \code{\link{is.ses}};
\code{\link{is.stat3D}}; \code{\link{is.stat}};
\code{\link{is.tdr3D}}; \code{\link{is.tdr}}
}

