\name{is.stat}
\alias{is.stat}
\title{is.stat}
\usage{
is.stat(x)
}
\arguments{
  \item{x}{The object}
}
\description{
Test if an object belong to the 'stat' class
}
\seealso{
Other is.thing: \code{\link{is.SES}}; \code{\link{is.ses}};
\code{\link{is.stat3D}}; \code{\link{is.tdr3D}};
\code{\link{is.tdr}}
}

