\name{is.stat3D}
\alias{is.stat3D}
\title{is.stat3D}
\usage{
is.stat3D(x)
}
\arguments{
  \item{x}{The object}
}
\description{
Test if an object belong to the 'stat3D' class
}
\seealso{
Other is.thing: \code{\link{is.SES}}; \code{\link{is.ses}};
\code{\link{is.stat}}; \code{\link{is.tdr3D}};
\code{\link{is.tdr}}
}

