\name{is.tdr}
\alias{is.tdr}
\title{is.tdr}
\usage{
  is.tdr(x)
}
\arguments{
  \item{x}{The object}
}
\description{
  Test if an object belong to the 'tdr' class
}
\seealso{
  Other is.thing: \code{\link{is.SES}},
  \code{\link{is.ses}}, \code{\link{is.stat}},
  \code{\link{is.stat3D}}, \code{\link{is.tdr3D}}
}

