\name{is.tdr3D}
\alias{is.tdr3D}
\title{is.tdr3D}
\usage{
  is.tdr3D(x)
}
\arguments{
  \item{x}{The object}
}
\description{
  Test if an object belong to the 'tdr3D' class
}
\seealso{
  Other is.thing: \code{\link{is.SES}},
  \code{\link{is.ses}}, \code{\link{is.stat}},
  \code{\link{is.stat3D}}, \code{\link{is.tdr}}
}

