\name{mZeu2layer}
\alias{mZeu2layer}
\title{mZeu2layer}
\usage{
  mZeu2layer(mZeu)
}
\arguments{
  \item{mZeu}{}
}
\description{
  Assign to a record the SEAPODYM layer it beyond using the
  depth expressed in multiple of Zeu.
}
\details{
  Layers limits taken to be 1.5, 4.5 and 10.5 time Zeu.
}
\author{
  Yves
}

