\name{nval}
\alias{nval}
\title{nval}
\usage{
  nval(x)
}
\arguments{
  \item{x}{The object to compute the number of distinct
  values.}
}
\description{
  Count the number of distinct values in an atomic vector.
}
\examples{
nval(rep(1:5, 5:1)) # 5
}
\seealso{
  Other generalUtils: \code{\link{SESname}},
  \code{\link{convertTime}}, \code{\link{per}},
  \code{\link{replaceMissing}}, \code{\link{sunPos}},
  \code{\link{sunPosition}}
}

