\name{SESplot}
\alias{SESplot}
\title{SESplot}
\usage{
  SESplot(ses, colorvar = NULL, isobath = NULL, ...)
}
\arguments{
  \item{ses}{An ses object (generated by importSES).}

  \item{...}{Other parameters to be passed to methods.}

  \item{colorvar}{A 'statdives' variable to plot along the
  trajectory.}

  \item{isobath}{Isobaths to draw. Default is NULL (none).
  Isobath at -1000 m corresponds approximatively to the
  Kerguelen shell boarder. Slow down the function
  (shapefile loading). Available isobaths range from -200
  to -4400 meters by 200 m.}
}
\description{
  To draw relevant chart of a ses object. Includes
  bathymetry taken from
  \url{https://www.ga.gov.au/products/servlet/controller?event=GEOCAT_DETAILS&catno=71552}.
  Available isobaths range from -200 to -4400 meters by 200
  m.
}
\examples{
\dontrun{
require(maps)
path <- system.file("extdata", package="SES")
pathname <- file.path(path, "2011-16_SES_example_accelero.mat")
ses <- importSES(pathname)
SESplot(ses, isobath=-1000, colorvar=ses$stat$Catch.numb)
library("RColorBrewer")
mycol <- brewer.pal(n=10, name="BrBG")
SESplot(ses, ses$stat$Dive.dur, pts.args=list(pch=""), implt.args=list(col=mycol), img.args=list(nx=150, ny=150))
}
}
\seealso{
  Other SESplot: \code{\link{SESplot.ses}}
}

