\name{SESplot.statdives}
\alias{SESplot.statdives}
\title{SESplot.statdives}
\usage{
  \method{SESplot}{ses} (obj, colorvar = NULL, cond = NULL,
    isobath = NULL, pts.args = list(), map.args = list(),
    img.args = list(), implt.args = list(),
    plt.args = list(), colorvarname)
}
\arguments{
  \item{obj}{An ses object (generated by importSES).}

  \item{colorvar}{A 'statdives' variable to plot along the
  trajectory.}

  \item{cond}{A logical vector giving how to subset the
  rows of the 'statdives' object.}

  \item{isobath}{Isobaths to draw. Default is NULL (none).
  Isobath at -1000 m corresponds approximatively to the
  Kerguelen shell boarder. Slow down the function
  (shapefile loading). Available isobaths range from -200
  to -4400 meters by 200 m.}

  \item{pts.args}{Other parameters to be passed to
  \code{\link{points}}. Defaults: \code{list(pch=19,
  cex=.1)}.}

  \item{map.args}{Other parameters to be passed to
  \code{\link{map}}. Defaults: \code{list(fill=TRUE,
  col="gray")}.}

  \item{img.args}{Other parameters to be passed to
  \code{\link{as.image}}. Notice the possibility to set
  \code{nx} and \code{ny} arguments.}

  \item{implt.args}{Other parameters to be passed to
  \code{\link{image.plot}}. Defaults:
  \code{list(legend.width=1, legend.mar=7.8,
  legend.shrink=.7,
  legend.lab=deparse(substitute(colorvar)),
  legend.line=3.5)} with
  \code{par(mar=c(4.2,4.2,3.5,7.9))}. Notice the
  possibility to set \code{zlim}, \code{col}, \code{breaks}
  arguments.}

  \item{plt.args}{Other parameters to be passed to
  \code{\link{plot}}. Notice the possibility to set
  \code{xlim} and \code{ylim} arguments.}

  \item{colorvarname}{Name for the legend. For
  compatibility when called from SESplot.ses}
}
\description{
  A S3 method for \code{SESplot()} with regular ses
  objects. Includes bathymetry taken from
  \url{https://www.ga.gov.au/products/servlet/controller?event=GEOCAT_DETAILS&catno=71552}.
}
\details{
  If this error show up 'data set ‘worldMapEnv’ not found'
  try \code{require(maps)}. See examples in
  \code{\link{SESplot}}.
}
\seealso{
  Other SESplot: \code{\link{SESplot}},
  \code{\link{SESplot.ses}}, \code{\link{SESplot.tdr3D}}
}

