\name{SESplot.tdr3D}
\alias{SESplot.tdr3D}
\title{SESplot.tdr3D}
\usage{
  SESplot.tdr3D(ses, pch = 19, cex = 1, Z = NULL,
    Z.col = tim.colors(64), ...)
}
\arguments{
  \item{ses}{An tdr3D object (generated by importSES).}

  \item{pch}{Plotting character.}

  \item{cex}{Plotting character size.}

  \item{Z}{A 'statdives' variable to plot along the
  trajectory.}

  \item{Z.col}{Colors for Z values.}

  \item{...}{Other parameters to be passed to
  \code{scatterplot3d()} or \code{par()}.}
}
\description{
  A method for \code{SESplot()} with tdr3D objects
}
\author{
  Yves
}

