\name{statFuns}
\alias{statFuns}
\title{statFuns}
\usage{
statFuns(type = c("double", "integer", "logical", "factor", "character"))
}
\arguments{
  \item{type}{The type of computation. To choose in
  \code{c("double", "integer", "logical", "factor",
  "character")}.}
}
\description{
Functions to apply to compute summary statistics given a
type of vector.
}
\keyword{internal}

