\name{summary.statdives}
\alias{summary.ses}
\alias{summary.statdives}
\title{summary.statdives}
\usage{
summary.statdives(object, na.rm = TRUE, digits = 2)
}
\arguments{
  \item{object}{An object for which a summary is desired.}

  \item{na.rm}{A logical value indicating whether \code{NA}
  values should be stripped before the computation
  proceeds.}

  \item{digits}{The number of digits to show in dive
  statistics.}
}
\description{
S3 method for 'statdives' objects.
}

