\name{summary.tdr}
\alias{summary.ses}
\alias{summary.tdr}
\title{summary.tdr}
\usage{
summary.tdr(object, na.rm = TRUE, complete = TRUE, digits = 2)
}
\arguments{
  \item{complete}{Should the function compute all
  statistics.}

  \item{object}{An object for which a summary is desired.}

  \item{na.rm}{A logical value indicating whether \code{NA}
  values should be stripped before the computation
  proceeds.}

  \item{digits}{The number of digits to show in dive
  statistics.}
}
\description{
S3 method for 'tdr' objects.
}

