\name{timeReso}
\alias{timeReso}
\title{Get the sampling rate of a TDR dataset}
\usage{
timeReso(obj, n = 10, type = c("period", "frequence"))
}
\arguments{
  \item{obj}{A 'ses' or 'tdr' object}

  \item{n}{the number of replicates to do.}

  \item{type}{Should the result be returned in seconds
  ('period') or hertz ('frequence') ?}
}
\description{
Get the sampling rate of a TDR dataset
}

