\name{userHeader}
\alias{userHeader}
\title{userHeader}
\usage{
userHeader(header, type)
}
\arguments{
  \item{header}{The default header.}

  \item{type}{The type of the object.}
}
\description{
Get the user headers associated with default headers of one
of a package object.
}
\seealso{
Other settings: \code{\link{importSES}};
\code{\link{print.fmtSES}}; \code{\link{renames}};
\code{\link{resetFmtSES}}; \code{\link{saveFmtSES}};
\code{\link{whichSESobj}}
}
\keyword{internal}

