\name{ymd2posx}
\alias{ymd2posx}
\title{Convert a date matrix or data frame (yyyy|mm|dd) to a POSIXct object.}
\usage{
ymd2posx(ymd, width = ncol(ymd))
}
\arguments{
  \item{ymd}{}

  \item{width}{The desired number of columns desired in the
  data frame.}
}
\description{
Convert a date matrix or data frame (yyyy|mm|dd) to a
POSIXct object.
}
\keyword{internal}

