\docType{data}
\name{formatSES}
\alias{formatSES}
\title{\code{importSES()}: User preferences}
\format{Four data frames, one for each of TDR and statdives datasets (classic or 3D), storing the names to give to the initial matlab variable (alias), whether or not they should be imported from the .mat file (keep) and a brief decription of them.}
\description{
  User preferences about SES formatting on import. To edit,
  simply type \code{formatSES} (or \code{formatSES$tdr}
  etc...) on console.
}
\author{
  Yves
}
\seealso{
  \code{\link{importSES}}
}
\keyword{datasets}

