\name{is.ses}
\alias{is.ses}
\title{is.ses}
\usage{
is.ses(x)
}
\arguments{
  \item{x}{The object}
}
\description{
Test if an object belong to the 'ses' class
}
\seealso{
Other is.thing: \code{\link{is.SES}};
\code{\link{is.stat3D}}; \code{\link{is.stat}};
\code{\link{is.tdr3D}}; \code{\link{is.tdr}}
}

