\name{modelMorel}
\alias{modelMorel}
\title{modelMorel}
\usage{
  modelMorel(chl, stat = NULL, append = TRUE)
}
\arguments{
  \item{chl}{}

  \item{stat}{A statdives object. 'chl' argument is
  ignored. The values are taken from this statdive object.}

  \item{append}{Should the variable be returned with the
  entire statdives object ?}
}
\description{
  Predict euphotic depth (m) from the surface chlorophyll
  concentration
}
\author{
  Yves
}
\references{
  Morel and Berthon (1989)
}

