\name{mZeu2layer}
\alias{mZeu2layer}
\title{mZeu2layer}
\usage{
  mZeu2layer(mZeu)
}
\arguments{
  \item{mZeu}{The depth expressed as multiple of the
  euphotic depth.}
}
\description{
  Assign to a record the SEAPODYM layer it belongs ond
  using the depth expressed in multiple of Zeu.7
}
\details{
  Layers limits taken to be 1.5, 4.5 and 10.5 time Zeu.
}
\seealso{
  Other chl: \code{\link{extractChl}},
  \code{\link{importChl}}, \code{\link{modelMorel}}
}

