\name{posx2ymd}
\alias{posx2ymd}
\title{Convert a POSIX object to a data frame with the desired number of columns.}
\usage{
posx2ymd(posx, width = 6)
}
\arguments{
  \item{posx}{The POSIX object to convert}

  \item{width}{The desired number of columns desired in the
  data frame.}
}
\description{
Convert a POSIX object to a data frame with the desired
number of columns.
}
\keyword{internal}

