\name{renames}
\alias{renames}
\title{renames}
\usage{
  renames(type = c("tdr", "stat"), obj, objtxt)
}
\arguments{
  \item{type}{The type of dataset}

  \item{obj}{The object to process.}

  \item{objtxt}{The original names. tdrcortxt or
  statdivestxt in the .mat file.}
}
\description{
  Extract and rename tdr and statdives variables according
  to formatSES instructions.
}
\author{
  Yves
}

