\name{SESplot}
\alias{SESplot}
\title{SESplot}
\usage{
  SESplot(ses, ...)
}
\arguments{
  \item{ses}{An ses object (generated by importSES).}

  \item{pch}{Plotting character.}

  \item{cex}{Plotting character size.}

  \item{fill}{Should the terrestrial area be filled with
  color.}

  \item{m.col}{The color to use in order to fill the
  terrestrial area.}

  \item{isobath}{Isobaths to draw. Default is NULL (none).
  Isobath at -1000 m corresponds approximatively to the
  Kerguelen shell boarder. Slow down the function
  (shapefile loading).}

  \item{Z}{A 'statdives' variable to plot along the
  trajectory.}

  \item{Z.col}{Colors for Z values.}

  \item{...}{Other parameters to be passed to \code{plot()}
  or \code{par()}.}
}
\description{
  To draw relevant chart of a ses object.
}
\details{
  Bathymetry taken from
  \url{https://www.ga.gov.au/products/servlet/controller?event=GEOCAT_DETAILS&catno=71552}.
  Isobaths from -200 to -4400 meters by 200 m.
}
\examples{
path <- system.file("extdata", package="SES")
pathname <- file.path(path, "2011-16_SES_example_accelero.mat")
ses <- importSES(pathname)
SESplot(ses, isobath=-1000, Z=ses$stat$Catch.numb)
\\notrun{
library("RColorBrewer")
mycol <- brewer.pal(n=10, name="BrBG")
SESplot(ses, Z=ses$stat$Dive.dur, Z.col=mycol)
}
}
\author{
  Yves
}

