\name{SESplot}
\alias{SESplot}
\title{SESplot}
\usage{
SESplot(ses, ...)
}
\arguments{
  \item{ses}{An ses object (generated by importSES).}

  \item{...}{Other parameters to be passed to methods.}
}
\description{
To draw relevant chart of a ses object. Includes bathymetry
taken from
\url{https://www.ga.gov.au/products/servlet/controller?event=GEOCAT_DETAILS&catno=71552}.
Available isobaths range from -200 to -4400 meters by 200
m.
}
\examples{
\dontrun{
path <- system.file("extdata", package="SES")
pathname <- file.path(path, "2011-16_SES_example_accelero.mat")
ses <- importSES(pathname)
SESplot(ses, isobath=-1000, Z=ses$stat$Catch.numb)
library("RColorBrewer")
mycol <- brewer.pal(n=10, name="BrBG")
SESplot(ses, Z=ses$stat$Dive.dur, Z.col=mycol)
}
}
\seealso{
Other SESplot: \code{\link{SESplot.ses}}
}

