\name{SESplot.ses}
\alias{SESplot.ses}
\title{SESplot.ses}
\usage{
\method{SESplot}{ses}(ses, pch = 19, cex = 1, fill = TRUE,
  m.col = "gray", isobath = NULL, Z = NULL, Z.col = tim.colors(64), ...)
}
\arguments{
  \item{ses}{An ses object (generated by importSES).}

  \item{pch}{Plotting character.}

  \item{cex}{Plotting character size.}

  \item{fill}{Should the terrestrial area be filled with
  color.}

  \item{m.col}{The color to use in order to fill the
  terrestrial area.}

  \item{isobath}{Isobaths to draw. Default is NULL (none).
  Isobath at -1000 m corresponds approximatively to the
  Kerguelen shell boarder. Slow down the function
  (shapefile loading).}

  \item{Z}{A 'statdives' variable to plot along the
  trajectory.}

  \item{Z.col}{Colors for Z values.}

  \item{...}{Other parameters to be passed to \code{plot()}
  or \code{par()}.}
}
\description{
A S3 method for \code{SESplot()} with regular ses objects.
Includes bathymetry taken from
\url{https://www.ga.gov.au/products/servlet/controller?event=GEOCAT_DETAILS&catno=71552}.
Available isobaths range from -200 to -4400 meters by 200
m.
}
\seealso{
Other SESplot: \code{\link{SESplot}}
}

