\name{summary.tdr}
\alias{summary.tdr}
\title{summary.tdr}
\usage{
  summary.tdr(object, na.rm = TRUE)
}
\arguments{
  \item{object}{An object for which a summary is desired.}

  \item{na.rm}{A logical value indicating whether \code{NA}
  values should be stripped before the computation
  proceeds.}
}
\description{
  S3 method for 'tdr' objects.
}

