\name{userHeader}
\alias{userHeader}
\title{userHeader}
\usage{
  userHeader(header, type)
}
\arguments{
  \item{header}{The default header.}

  \item{type}{The type of the object.}
}
\description{
  Get the user header associated with a default header in
  an object.
}
\author{
  Yves
}
\keyword{internal}

