\name{whichSESobj}
\alias{whichSESobj}
\title{whichSESobj}
\usage{
whichSESobj(x)
}
\arguments{
  \item{x}{}
}
\description{
Return the name of formatSES element matching x.
}
\seealso{
Other settings: \code{\link{importSES}};
\code{\link{print.fmtSES}}; \code{\link{renames}};
\code{\link{resetFmtSES}}; \code{\link{saveFmtSES}};
\code{\link{userHeader}}
}
\keyword{internal}

